Add global_det atmos prep restart
Refs: #1

+++ scripts/global_det/prep/exevs_global_det_atmos_prep.sh
@@ -13,25 +13,15 @@ echo
 ############################################################
 ## Global Deterministic Atmospheric Prep
 ############################################################
+python ${USHevs}/global_det/global_det_atmos_restart.py
+status=$?
+[[ $status -ne 0 ]] && exit $status
+[[ $status -eq 0 ]] && echo "Succesfully ran ${USHevs}/global_det/global_det_atmos_restart.py"
+
 python ${USHevs}/global_det/global_det_atmos_prep.py
 status=$?
 [[ $status -ne 0 ]] && exit $status
 [[ $status -eq 0 ]] && echo "Succesfully ran ${USHevs}/global_det/global_det_atmos_prep.py"
-
-if [ $SENDCOM = YES ]; then
-    for mname in $MODELNAME; do
-        mkdir -p $COMOUT.$INITDATE/$mname
-        for FILE in $DATA/$RUN.$INITDATE/$mname/*; do
-            cp -v $FILE $COMOUT.$INITDATE/$mname/.
-        done
-    done
-    for oname in $OBSNAME; do
-        mkdir -p $COMOUT.$INITDATE/$oname
-        for FILE in $DATA/$RUN.$INITDATE/$oname/*; do
-            cp -v $FILE $COMOUT.$INITDATE/$oname/.
-        done
-    done
-fi
 
 if [ ${KEEPDATA} != YES ]; then
     rm -rf $DATA
